Rename ConversingUsers to Chatters.
