+++ src/Chatters.js
@@ -1,5 +1,3 @@
-// TODO: Rename this file to Chatters.js
-// Will do it in its own commit so as to not complicate the CRs.
 const { Map } = require("immutable");
 
 const chatterIds = Object.freeze({
