+++ src/Chatters.test.js
@@ -1,4 +1,4 @@
-import { chatters, chatterIds } from "./ConversingUsers";
+import { chatters, chatterIds } from "./Chatters";
 
 test("that all userIds are contained in the map and that all users in the map are represented in the userIds", () => {
   expect(Object.keys(chatterIds).length).toEqual(
